Initial Commit
This is the first commit to the sdig server.
Sdig is the server side of a my personal messaging app "batcher".
As of now there is no client side, but hopefully I will start making the
it after finishing this semester.

Batcher will be a group chat only messaging app, currently I plan to
make it text only, but if I gained more experiance in the future, I
might consider adding image and video support.

Calling seems too complicated so I will never add it, I might add voice
messages though.

Currently I am working on this project alone, but I might get other
people to help me in this project, especially with the client side, as I
have no experiance.

Note that the current name "Batcher" is mostly temporary and might
change in the future.