Add comments and change multiple things
I added comments that describe what each struct does and the purpose of
each member alogside more information.
I also did the same to functions even though they are that descriptive.

I changed how the login works and made the chat manager handle the login
instead of the user handle function because I didn't want to have a
shared mutex in every single user.

I also wrote comments that list what to do in case I forgot.

I added a way to send a message(not a chat message) to the client to
either notify them or send them an error or a warning