Handle the requests from the user to the chat manager
I added the handling of all chat manager requests.

Now the client can make a new user account or delete it.
Log in an Log out.
Create a new chat and join it as the owner of the chat, and delete the
chat if the request is from the owner.
Join the chat of another user and send messages.